Move StoreItem.kt to other package.
