+++ app/src/main/java/com/shov/unlimstorage/models/filesRepository/FilesInteractor.kt
@@ -1,6 +1,8 @@
 package com.shov.unlimstorage.models.filesRepository
 
-import com.shov.unlimstorage.models.StoreItem
+import com.shov.unlimstorage.models.items.ItemType
+import com.shov.unlimstorage.models.items.StoreItem
+import com.shov.unlimstorage.models.items.StoreMetadataItem
 
 interface FilesInteractor {
 	fun getFiles(folderId: String? = null): List<StoreItem>
