+++ app/src/main/java/com/shov/unlimstorage/models/filesRepository/FilesRepository.kt
@@ -1,7 +1,9 @@
 package com.shov.unlimstorage.models.filesRepository
 
 import com.shov.unlimstorage.db.StoreItemDao
-import com.shov.unlimstorage.models.StoreItem
+import com.shov.unlimstorage.models.items.ItemType
+import com.shov.unlimstorage.models.items.StoreItem
+import com.shov.unlimstorage.models.items.StoreMetadataItem
 import com.shov.unlimstorage.models.signInModels.StorageType
 import javax.inject.Inject
